# docs/swagger.yaml
basePath: /api
definitions:
...
  icmongolang_internal_users_presenter.UserCreate:
    properties:
      confirm_password:
        example: password
        minLength: 8
        type: string
      email:
        example: kongnakornna@gmail.com
        type: string
      firstname:
        example: Kongnakorn
        type: string
      fullname:
        example: Kongnakorn Jantakun
        type: string
      lastname:
        example: Jantakun
        type: string
      line_id:
        example: kongnakorn_line
        type: string
      location_id:
        example: loc_001
        type: string
      mobile_number:
        example: "0812345678"
        type: string
      password:
        example: password
        minLength: 8
        type: string
      phone_number:
        example: "021234567"
        type: string
      role_id:
        example: 2
        minimum: 1
        type: integer
    required:
    - confirm_password
    - email
    - password
    - role_id
    type: object
  icmongolang_internal_users_presenter.UserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      firstname:
        type: string
      fullname:
        type: string
      id:
        type: string
      is_superuser:
        type: boolean
      lastname:
        type: string
      line_id:
        type: string
      location_id:
        type: string
      mobile_number:
        type: string
      phone_number:
        type: string
      role_id:
        type: integer
      status:
        description: 1 active, 0 inactive
        type: integer
      updated_at:
        type: string
      verified:
        type: boolean
    type: object
  icmongolang_internal_users_presenter.UserUpdate:
    properties:
      firstname:
        type: string
      fullname:
        type: string
      lastname:
        type: string
      line_id:
        type: string
      location_id:
        type: string
      mobile_number:
        type: string
      phone_number:
        type: string
    type: object
  icmongolang_internal_users_presenter.UserUpdatePassword:
    properties:
      confirm_password:
        minLength: 8
        type: string
      new_password:
        minLength: 8
        type: string
      old_password:
        minLength: 8
        type: string
    required:
    - confirm_password
    - new_password
    - old_password
    type: object
...
  responses.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/httpErrors.ErrResponse'
      is_success:
        example: false
        type: boolean
    type: object
  responses.SuccessResponse-array_icmongolang_internal_users_presenter_UserResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/icmongolang_internal_users_presenter.UserResponse'
        type: array
      is_success:
        example: true
        type: boolean
    type: object
...
  responses.SuccessResponse-icmongolang_internal_users_presenter_UserResponse:
    properties:
      data:
        $ref: '#/definitions/icmongolang_internal_users_presenter.UserResponse'
      is_success:
        example: true
        type: boolean
    type: object
...
info:
  contact: {}
  description: API for ICMON project
  title: Go Rest API
  version: "1.0"
paths:
...
  /user:
    get:
      consumes:
      - application/json
      description: Retrieve users.
      parameters:
      - description: limit
        format: limit
        in: query
        name: limit
        type: integer
      - description: offset
        format: offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.SuccessResponse-array_icmongolang_internal_users_presenter_UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - OAuth2Password: []
      summary: Read Users
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Create new user.
      parameters:
      - description: Add user
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/icmongolang_internal_users_presenter.UserCreate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.SuccessResponse-icmongolang_internal_users_presenter_UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - OAuth2Password: []
      summary: Create User
      tags:
      - users
  /user/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an user by ID.
      parameters:
      - description: User Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.SuccessResponse-icmongolang_internal_users_presenter_UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - OAuth2Password: []
      summary: Delete user
      tags:
      - users
    get:
      consumes:
      - application/json
      description: Get user by ID.
      parameters:
      - description: User Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.SuccessResponse-icmongolang_internal_users_presenter_UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - OAuth2Password: []
      summary: Read user
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update an user by ID.
      parameters:
      - description: User Id
        in: path
        name: id
        required: true
        type: string
      - description: Update user
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/icmongolang_internal_users_presenter.UserUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.SuccessResponse-icmongolang_internal_users_presenter_UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - OAuth2Password: []
      summary: Update user
      tags:
      - users
  /user/{id}/logoutall:
    get:
      consumes:
      - application/json
      description: Logout all session of user with id.
      parameters:
      - description: User Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - OAuth2Password: []
      summary: Logout all of user
      tags:
      - users
  /user/{id}/updatepass:
    patch:
      consumes:
      - application/json
      description: Update password user by ID.
      parameters:
      - description: User Id
        in: path
        name: id
        required: true
        type: string
      - description: Update user
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/icmongolang_internal_users_presenter.UserUpdatePassword'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.SuccessResponse-icmongolang_internal_users_presenter_UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - OAuth2Password: []
      summary: Update password user
      tags:
      - users
  /user/me:
    get:
      consumes:
      - application/json
      description: Get user me.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.SuccessResponse-icmongolang_internal_users_presenter_UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - OAuth2Password: []
      summary: Read user me
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update user me.
      parameters:
      - description: Update user
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/icmongolang_internal_users_presenter.UserUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.SuccessResponse-icmongolang_internal_users_presenter_UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - OAuth2Password: []
      summary: Update user me
      tags:
      - users
  /user/me/updatepass:
    patch:
      consumes:
      - application/json
      description: Update password user me.
      parameters:
      - description: Update user
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/icmongolang_internal_users_presenter.UserUpdatePassword'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.SuccessResponse-icmongolang_internal_users_presenter_UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - OAuth2Password: []
      summary: Update password user me
      tags:
      - users
securityDefinitions:
  OAuth2Password:
    flow: password
    scopes:
      read: Grants read access
      write: Grants write access
    tokenUrl: /api/auth/login
    type: oauth2
swagger: "2.0"